requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Django==3.1.4
django-pandas==0.6.2
djangorestframework==3.12.2
idna==2.10
libarchive-c==2.9
lxml==4.6.2
nose==1.3.7
numpy==1.19.4
pandas==1.1.5
Pillow==8.1.0
psycopg2-binary==2.8.6
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
tqdm==4.55.1
urllib3==1.26.2
webdriver-manager==3.2.2
